Add files via upload
Updated code for the date format is consistent (part c in output)! Also changed part b for output portion because requirements include having the csv file ordered by ID number. However, the new method used is only good for running the code once and not more than that because it doesn't rewrite the file at all, it just makes a new file (if file doesn't already exist) and appends (if the file already exists)